move components to vc folder
